+++ BSEAV/app/atlas/src/widgets/widget_popup_list.h
@@ -1,7 +1,7 @@
-/***************************************************************************
- * (c) 2002-2015 Broadcom Corporation
+/******************************************************************************
+ * Broadcom Proprietary and Confidential. (c)2016 Broadcom. All rights reserved.
  *
- * This program is the proprietary software of Broadcom Corporation and/or its
+ * This program is the proprietary software of Broadcom and/or its
  * licensors, and may only be used, duplicated, modified or distributed pursuant
  * to the terms and conditions of a separate, written license agreement executed
  * between you and Broadcom (an "Authorized License").  Except as set forth in
@@ -37,7 +37,6 @@
  *    OF THE AMOUNT ACTUALLY PAID FOR THE SOFTWARE ITSELF OR U.S. $1, WHICHEVER
  *    IS GREATER. THESE LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF
  *    ESSENTIAL PURPOSE OF ANY LIMITED REMEDY.
- *
  *****************************************************************************/
 
 #ifndef ATLAS_WIDGET_POPUP_LIST_H__
@@ -71,7 +70,7 @@ public:
     void                 layout(void);
     void                 popup(bool show);
     CWidgetCheckButton * addButton(const char * name, uint16_t width, uint16_t height);
-    CWidgetCheckButton * findButton(int value);
+    CWidgetCheckButton * findButton(long value);
     void                 setActive(
             CWidgetButton * pWidgetItem,
             bool            bActive
@@ -80,7 +79,7 @@ public:
     void                 clearButtons(void);
     eRet                 select(CWidgetCheckButton * pButton);
     eRet                 select(const char * name);
-    eRet                 select(const int value);
+    eRet                 select(const long value);
     CWidgetCheckButton * getSelection(void) { return(_pSelectedButton); }
     void                 onClick(bwidget_t widget);
     eRet                 sort(void) { return(_pListView->sort()); }
