+++ nexus/examples/multiprocess/client_stress_test.c
@@ -1,7 +1,7 @@
 /******************************************************************************
- *    (c)2008-2013 Broadcom Corporation
+ *  Broadcom Proprietary and Confidential. (c)2008-2016 Broadcom. All rights reserved.
  *
- * This program is the proprietary software of Broadcom Corporation and/or its licensors,
+ * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
  * conditions of a separate, written license agreement executed between you and Broadcom
  * (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
@@ -173,7 +173,7 @@ int main(void)
 
     for (i=0;i<20;i++) {
         NEXUS_Graphics2D_GetDefaultFillSettings(&fillSettings);
-        fillSettings.surface = (void*)rand();
+        fillSettings.surface = (void*)(long)rand();
         if (fillSettings.surface == surface) continue; /* just in case we get lucky */
         fillSettings.rect.width = rand();
         fillSettings.rect.height = rand();
