+++ nexus/examples/multiprocess/input_router.c
@@ -1,7 +1,7 @@
 /******************************************************************************
- *    (c)2008-2013 Broadcom Corporation
+ * Broadcom Proprietary and Confidential. (c)2016 Broadcom. All rights reserved.
  *
- * This program is the proprietary software of Broadcom Corporation and/or its licensors,
+ * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
  * conditions of a separate, written license agreement executed between you and Broadcom
  * (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
@@ -35,16 +35,8 @@
  * LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF ESSENTIAL PURPOSE OF
  * ANY LIMITED REMEDY.
  *
- * $brcm_Workfile: $
- * $brcm_Revision: $
- * $brcm_Date: $
- *
  * Module Description:
  *
- * Revision History:
- *
- * $brcm_Log: $
- * 
  *****************************************************************************/
 
 #include "nexus_platform.h"
@@ -141,7 +133,7 @@ int main(void)
         
         if (irInput) {
             NEXUS_IrInputEvent irEvent;
-            unsigned n;
+            size_t n;
             bool overflow;
             rc = NEXUS_IrInput_GetEvents(irInput, &irEvent, 1, &n, &overflow);
             if (n == 1) {
@@ -160,7 +152,7 @@ int main(void)
 #if NEXUS_HAS_KEYPAD
         if (keypad) {
             NEXUS_KeypadEvent keypadEvent;
-            unsigned n;
+            size_t n;
             bool overflow;
             rc = NEXUS_Keypad_GetEvents(keypad, &keypadEvent, 1, &n, &overflow);
             BDBG_ASSERT(!rc);
